rpm: find-legacy-runtime.sh: Use .L suffix for better upgrades.
This makes the legacy package provided by the script always newer
than the original. This improves identification and, together
with a more strict Obsoltes, will cause this new legacy package
to replace some older packages that would not be replaced
otherwise (this also covers a possible case when the new legacy
package happens to have the exact same name as the original one).



git-svn-id: http://svn.netlabs.org/repos/rpm/rpm/trunk@984 978522c7-42b7-df11-88a9-00e081727c95